+++ templates/Vector.cpp
@@ -21,13 +21,13 @@ Vector<T>::~Vector(void)
  */
 
 template <typename T>
-unsigned int	Vector<T>::size(void) const
+unsigned long	Vector<T>::size(void) const
 {
 	return this->_size;
 }
 
 template <typename T>
-unsigned int	Vector<T>::max_size(void) const
+unsigned long	Vector<T>::max_size(void) const
 {
 	return this->_allocator.max_size();
 }
@@ -43,13 +43,13 @@ bool	Vector<T>::empty(void) const
  */
 
 template <typename T>
-T	&Vector<T>::operator[](unsigned int position) const
+T	&Vector<T>::operator[](unsigned long position) const
 {
 	return this->_list[position];
 }
 
 template <typename T>
-T	&Vector<T>::at(unsigned int position) const
+T	&Vector<T>::at(unsigned long position) const
 {
 	if (this->size() <= position)
 		throw out_of_range();
@@ -73,7 +73,7 @@ T	&Vector<T>::back(void) const
  */
 
 template <typename T>
-void	Vector<T>::assign(unsigned int nElements, T value)
+void	Vector<T>::assign(unsigned long nElements, T value)
 {
 	message("Vector: assign");
 
@@ -84,7 +84,7 @@ void	Vector<T>::assign(unsigned int nElements, T value)
 	this->_maxSize = nElements;
 	this->_list = _allocator.allocate(nElements);
 
-	for (unsigned int i = 0; i < nElements; i++)
+	for (unsigned long i = 0; i < nElements; i++)
 		this->_list[i] = value;
 }
 
@@ -123,7 +123,7 @@ template <typename T>
 std::ostream	&operator<<(std::ostream &output, Vector<T> const & vec)
 {
 	output << "(vector) = size=" << vec.size() << " {" << std::endl;
-	for (unsigned int i = 0; i < vec.size(); i++)
+	for (unsigned long i = 0; i < vec.size(); i++)
 		output << "  ["<< i << "] = " << vec[i] << std::endl;
 	output << "}" << std::endl;
 	return output;
